feat/#36 : index.tsx 제거 및 내용 page.tsx로 이동:
