--- src/app/stage/_components/stageCarousel/index.tsx
@@ -1,18 +1,0 @@
-import React from "react";
-
-import EmblaCarousel from "./EmblaCarousel";
-import { EmblaOptionsType } from "embla-carousel";
-
-import "./embla.css";
-
-const OPTIONS: EmblaOptionsType = {};
-const SLIDE_COUNT = 3;
-const SLIDES = Array.from(Array(SLIDE_COUNT).keys());
-
-export const Carousel = () => (
-  <main className="sandbox">
-    <section className="sandbox__carousel">
-      <EmblaCarousel slides={SLIDES} options={OPTIONS} />
-    </section>
-  </main>
-);
